Display Updates
Added the ability to add a layout on top of the game that has buttons.
Buttons can now easily be added by adding them to the game_buttons.xml file and will automatically be added to the user interface for the game. However if new buttons are added they still must be added to the game logic